% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rchk.R
\name{read_rchk}
\alias{read_rchk}
\title{Read rchk output diagnostics into a tibble}
\usage{
read_rchk(file, ...)
}
\arguments{
\item{file, ...}{file and extract parameters to [readr::read_lines()]}
}
\value{
a tibble
}
\description{
Read rchk output diagnostics into a tibble
}
